+++ libr/anal/Makefile
@@ -32,7 +32,7 @@ OBJLIBS=meta.o reflines.o op.o fcn.o bb.o var.o block.o
 OBJLIBS+=cond.o value.o cc.o class.o diff.o type.o type_pdb.o dwarf_process.o
 OBJLIBS+=hint.o anal.o data.o xrefs.o sign.o
 OBJLIBS+=switch.o cycles.o esil_dfg.o esil_cfg.o
-OBJLIBS+=flirt.o labels.o
+OBJLIBS+=flirt.o labels.o cparse.o
 OBJLIBS+=pin.o vtable.o rtti.o codemeta.o anplugs.o global.o
 OBJLIBS+=rtti_msvc.o rtti_itanium.o jmptbl.o function.o
 
@@ -86,5 +86,15 @@ do:
 
 doclean:
 	$(MAKE) -C d clean
+
+tcc-sync sync-tcc:
+	rm -rf _
+	git clone --depth 1 git://repo.or.cz/tinycc.git _
+	cd _ ; ./configure --prefix=${PREFIX}
+	mkdir -p tcc
+	for FILE in ${CFILES} ${HFILES} ; do cp -f _/$$FILE tcc ; done
+	cp -f _/VERSION _/COPYING _/stab.def tcc
+	git add tcc
+	rm -rf _
 
 include $(LTOP)/rules.mk
