+++ libr/anal/cparse.c
@@ -4,7 +4,7 @@
 #include "r_types.h"
 #include "r_parse.h"
 #define ONE_SOURCE 1
-#include "tcc.h"
+#include "c/tcc.h"
 #include "c/tccgen.c"
 #include "c/tccpp.c"
 #include "c/libtcc.c"
@@ -119,7 +119,7 @@ static TCCState *new_tcc(RAnal *anal) {
 	return ts;
 }
 
-R_API char *r_parse_c_file(RAnal *anal, const char *path, const char *dir, char **error_msg) {
+R_API char *r_anal_cparse_file(RAnal *anal, const char *path, const char *dir, char **error_msg) {
 	char *str = NULL;
 	TCCState *s1 = new_tcc (anal);
 	if (!s1) {
@@ -152,7 +152,7 @@ R_API char *r_parse_c_file(RAnal *anal, const char *path, const char *dir, char 
 	return str;
 }
 
-R_API char *r_parse_c_string(RAnal *anal, const char *code, char **error_msg) {
+R_API char *r_anal_cparse(RAnal *anal, const char *code, char **error_msg) {
 	char *str = NULL;
 	TCCState *s1 = new_tcc (anal);
 	if (!s1) {
