+++ libr/parse/meson.build
@@ -1,25 +1,5 @@
 # handle tcc dependency (heavily modified version)
-tcc_files = [
-  'c/libtcc.c',
-  'c/tccgen.c',
-  'c/tccpp.c'
-]
-
-tcc_inc = [platform_inc, include_directories('c')]
-
-#libr2tcc = static_library('r2tcc', tcc_files,
-#  dependencies: [r_util_dep],
-#  include_directories: tcc_inc,
-#  implicit_include_directories: false
-#)
-
-tcc_dep = declare_dependency(
- #  link_whole: libr2tcc,
-  include_directories: tcc_inc
-)
-
 r_parse_sources = [
-  'code.c',
   'filter.c',
   'parse.c',
   'p/parse_6502_pseudo.c',
@@ -41,7 +21,6 @@ r_parse_sources = [
   'p/parse_riscv_pseudo.c',
   'p/parse_z80_pseudo.c'
 ]
-# r_parse_sources += tcc_files
 
 r_parse = library('r_parse', r_parse_sources,
   include_directories: platform_inc,
@@ -51,8 +30,7 @@ r_parse = library('r_parse', r_parse_sources,
     r_flag_dep,
     r_syscall_dep,
     r_reg_dep,
-    r_cons_dep,
-    tcc_dep
+    r_cons_dep
   ],
   install: true,
   implicit_include_directories: false,
@@ -72,8 +50,7 @@ r_parse_static = static_library('r_parse_static', r_parse_sources,
     r_flag_static_dep,
     r_syscall_static_dep,
     r_reg_static_dep,
-    r_cons_static_dep,
-    tcc_dep
+    r_cons_static_dep
   ],
   install: true,
   implicit_include_directories: false,
